actions are now swapped between both players
also, server_agent and client_agent dictionaries now use just the name of the agent instead of the full node for rpc-compatibility. Also, added in the starts of the aiming indicator, and code for removing both. Just need to save the actions to the agents prior to starting the execution phase.