massively boosted replication with pre-generation
weapons are now all generated as nodes, either immediately or one process frame later in the case of weapon pickups, when the match starts.
the names of weapons are used for interacting with their nodes, through GameRefs.get_weapon_node() and GameRefs.get_pickup_node(), necessitating a team's agents to all have different names, but this is a problem that I'll easily circumvent in Menu,gd.
as well, much code has been consolidated, trimmed, or otherwise simplified or easier replication (i.e. agents now store just the weapon names in held_weapons in order to replicate)